chore: refactor batch write modules
