+++ index.js
@@ -2,7 +2,7 @@ const { DynamoDB } = require('@aws-sdk/client-dynamodb');
 const { DynamoDBDocument } = require('@aws-sdk/lib-dynamodb');
 
 const getItem = require('./src/dynamodb/get-item');
-const batchWrite = require('./src/dynamodb/batch-write');
+const batchWrite = require('./src/dynamodb/batch/batch-write');
 
 function createDynamoClient(documentClient) {
   return {
