+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@mhlabs/dynamodb-client",
-  "version": "0.5.0",
+  "version": "0.6.0",
   "description": "A lib to simplify working with dynamodb documentclient in aws sdk v3.",
   "main": "index.js",
   "jest": {
