+++ src/dynamodb/batch/constants.js
@@ -1,0 +1,6 @@
+module.exports = {
+  MAX_ITEMS_PER_BATCH: 25,
+  UNPROCESSED_RETRY_LIMIT: 2,
+  DEFAULT_UNPROCESSED_MIN_RETRY_TIMOUT_MS: 300,
+  DEFAULT_UNPROCESSED_MAX_RETRY_TIMOUT_MS: 1000
+};
